requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0
urllib3==2.0.7
selenium==4.15.0
webdriver-manager==4.0.1
undetected-chromedriver==3.5.4
setuptools>=65.0.0
psycopg2-binary==2.9.9
